+++ lib/src/framework/draw/overlay.dart
@@ -15,6 +15,7 @@
 library;
 
 import 'dart:collection';
+import 'dart:math' as math;
 import 'package:flexi_formatter/flexi_formatter.dart';
 import 'package:flutter/widgets.dart';
 import 'package:json_annotation/json_annotation.dart';